sinosfera_prj/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.0.0
Django==4.2.4
django-admin-interface==0.26.1
django-admin-tools==0.9.3
django-allauth==0.54.0
django-appconf==1.0.5
django-baton==2.8.0
django-bootstrap-v5==1.0.11
django-colorfield==0.10.1
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-geojson==4.0.0
django-grappelli==3.0.7
django-import-export==3.2.0
django-leaflet==0.29.0
django-model-utils==4.3.1
django-project-version==0.18.1
django-slick-reporting==1.0.1
django-smart-selects==1.6.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
docutils==0.20.1
et-xmlfile==1.1.0
GDAL @ file:///D:/HDD_Code/SINOSFERA_Workspace/sinosfera_prj/GDAL-3.4.3-cp311-cp311-win_amd64.whl#sha256=f78861fb5115d5c2f8cf3c52a492ff548da9e1256dc84088947379f90e77e5b6
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
idna==3.4
isort==5.12.0
MarkupPy==1.14
MarkupSafe==2.1.3
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
pipdeptree==2.13.0
pretty-errors==1.2.25
protobuf==4.24.2
psycopg2==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==3.0.0
xlrd==2.0.1
xlwt==1.3.0
